a lot more robustness with a copy-on-write approach.
Now we ckeep track of a list of pods identified by namespace and name.
Instead of mutating a network data structure, we create the networks
from scratch so that we only need to implement Add and no Delete.
This also supports scenario's where annotations change at runtime and
will also make it easier to support multiple networks.